Implement complete EAFP architecture with explicit solver-to-interface mapping
## Major Improvements

### 1. Explicit Solver Routing
- Replace trial-and-error EAFP with explicit solver-to-interface mapping
- Eliminate unnecessary ValueError exceptions for cross-interface calls
- Direct routing: matlab_sedumi → MATLAB interface, cvxpy_ecos → Python interface

### 2. Complete Lazy Initialization
- Solver interfaces only initialize when actually accessed
- Remove module-level MATLAB availability checks
- Properties-based lazy initialization with proper error handling

### 3. Unified Interface Architecture
- Remove solver_registry.yaml (redundant with interface configurations)
- Single source of truth: interface *_SOLVER_CONFIGS
- Dynamic mapping generation from interface configurations

### 4. Performance Optimizations
- Zero unnecessary exceptions during normal operation
- O(1) solver lookup instead of trial-and-error
- Faster startup with lazy interface initialization

## Technical Changes

### Core Files Modified:
- scripts/benchmark/runner.py: Add explicit mapping and lazy properties
- scripts/solvers/python/python_interface.py: Implement lazy detection
- scripts/solvers/matlab_octave/matlab_interface.py: Implement lazy detection
- main.py: Fix import syntax error
- docs/development/detail_design.md: Update documentation

### Files Removed:
- config/solver_registry.yaml: Redundant with interface configurations

### Architecture Benefits:
- Eliminates: matlab_sedumi → Python (ValueError) → MATLAB flow
- Achieves: matlab_sedumi → mapping['matlab_sedumi'] = 'matlab' → MATLAB flow
- Clear error messages with available solver lists
- Extensible for future solver ecosystems (Julia, R, etc.)
