--- config/solver_registry.yaml
@@ -1,35 +1,0 @@
-# Simplified solver registry - only display names for reporting
-# Actual solver initialization logic is in code for better maintainability
-solvers:
-  scipy_linprog:
-    display_name: "SciPy linprog"
-    
-  cvxpy_clarabel:
-    display_name: "CLARABEL (via CVXPY)"
-    
-  cvxpy_scs:
-    display_name: "SCS (via CVXPY)"
-    
-  cvxpy_ecos:
-    display_name: "ECOS (via CVXPY)"
-    
-  cvxpy_osqp:
-    display_name: "OSQP (via CVXPY)"
-    
-  cvxpy_cvxopt:
-    display_name: "CVXOPT (via CVXPY)"
-    
-  cvxpy_sdpa:
-    display_name: "SDPA (via CVXPY)"
-    
-  cvxpy_scip:
-    display_name: "SCIP (via CVXPY)"
-    
-  cvxpy_highs:
-    display_name: "HiGHS (via CVXPY)"
-    
-  matlab_sedumi:
-    display_name: "SeDuMi (via MATLAB)"
-  
-  matlab_sdpt3:
-    display_name: "SDPT3 (via MATLAB)"
